Add a room route for users that simulates connecting to a sprint room.
In this phase, I would need to set up websockets for a joining queue, then 2 people who match in either skill or tech stacks would get out of the queue and be put into a sprint room. If users are joining a room like it is a chat-system then I would just need to set up the chat system within that roomId.