changed out code editing works due to some bugs.
Previously, if user1 were to create a new file and edit it then user2 would not have that file and also their App.js would be edited instead. Currently, the code_change takes into account the filename and the code content. I need a way to have the files appear as soon as one user makes them and not need to click any button to refresh it.